+++ haze_removal/test/main.cpp
@@ -21,7 +21,7 @@ int main(int argc, char* argv[])
     cv::Vec3b A;
     estimateAtmosphericLight(img, dark_channel, A);
 
-    cv::Mat_<float> t, t_refine; cv::Mat t_show;
+    cv::Mat_<double> t, t_refine; cv::Mat t_show;
     double min_value, max_value;
     initTransMap(img, A, t, 10);
     cv::Mat tMat = t;
@@ -33,14 +33,14 @@ int main(int argc, char* argv[])
     cv::Mat_<cv::Vec3b> recoverImg;
     recoverSceneRadiance(img, recoverImg, t_refine, A);
     //int size[] = {3, 3};
-    //cv::SparseMat_<float> CoeffMat(2, size);
+    //cv::SparseMat_<double> CoeffMat(2, size);
     //CoeffMat.ref(0, 0) = 4; CoeffMat.ref(0, 1) = 3;
     //CoeffMat.ref(1, 0) = 3; CoeffMat.ref(1, 1) = 4; CoeffMat.ref(1, 2) = -1;
     //CoeffMat.ref(2, 1) = -1; CoeffMat.ref(2, 2) = 4;
 
-    //cv::Mat_<float> b(3, 1);
+    //cv::Mat_<double> b(3, 1);
     //b(0, 0) = 24; b(1, 0) = 30; b(2, 0) = -24;
-    //cv::Mat_<float> X;
+    //cv::Mat_<double> X;
     //linearEquationSolver(CoeffMat, b, X);
 
     //std::cout << "X = " << X << std::endl;
